refactor: move number formatting functions to utils for reuse
